mcp-server/requirements.txt
fastmcp>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.1
openai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
